modified some files and add new file
